flask/app/requirements.txt
flask
flask-bootstrap
pylint
flask-moment
flask-wtf
flask-sqlalchemy
flask-migrate
flask-mail
flask-pymongo